servers/google_calendar/pyproject.toml
[project]
name = "orion-gcal-mcp"
version = "0.1.0"
description = "Orion Google Calendar MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "redis>=5.0.0",
    "arq>=0.26.0",
    "google-api-python-client>=2.128.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "mcp>=1.0.0",
    "pydantic>=2.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
